// source/enums.ts
export enum Sections {
  cards = "cards",
  deck = "deck",
  gallery = "gallery",
  nft = "nft",
}

export enum CardSuits {
  s = "spades",
  c = "clubs",
  h = "hearts",
  d = "diamonds",
  r = "red",
  b = "black",
}

export enum CardValues {
  j = "jack",
  q = "queen",
  k = "king",
  a = "ace",
}